dbops-related C# tests
